backend/app/core/context_agent/indian_sources.py
"""Additional MCP Source Fetchers - Economic Times, NSE, BSE

These fetchers provide Indian market-specific news and corporate announcements
with proper citation structure.
"""

import logging
import asyncio
import re
from typing import List, Dict, Any, Optional
from datetime import datetime, timedelta
import httpx
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class EconomicTimesMarketsFetcher:
    """
    Fetch market news and analysis from Economic Times Markets
    
    Economic Times is India's leading financial newspaper.
    This fetcher scrapes public market news related to specific stocks.
    """
    
    BASE_URL = "https://economictimes.indiatimes.com"
    SEARCH_URL = f"{BASE_URL}/markets/stocks/news"
    TIMEOUT_SECONDS = 10
    
    async def fetch_stock_news(
        self,
        ticker: str,
        company_name: Optional[str] = None,
        max_results: int = 5
    ) -> List[Dict[str, Any]]:
        """
        Fetch stock-specific news from Economic Times Markets
        
        Note: This is a placeholder implementation. In production:
        - Use ET's official API if available
        - Implement proper scraping with robust selectors
        - Add retry logic and rate limiting
        
        Args:
            ticker: Stock ticker (e.g., "RELIANCE", "TCS")
            company_name: Full company name for better search (optional)
            max_results: Maximum number of articles to return
        
        Returns:
            List of news items with title, url, published_at
        """
        news_items = []
        
        try:
            # Clean ticker (remove .NS/.BO suffix)
            clean_ticker = ticker.replace('.NS', '').replace('.BO', '')
            
            # For now, return empty list with logging
            # TODO: Implement actual ET Markets scraper or API integration
            logger.info(f"ET Markets fetch not yet implemented for {clean_ticker}")
            
            # Placeholder: Would fetch from ET Markets API/scraper here
            # async with httpx.AsyncClient(timeout=self.TIMEOUT_SECONDS) as client:
            #     response = await client.get(...)
            #     news_items = parse_et_response(response)
        
        except Exception as e:
            logger.debug(f"ET Markets fetch failed: {e}")
        
        return news_items


class NSEAnnouncementsFetcher:
    """
    Fetch corporate announcements from NSE India
    
    Uses NSE's public API for corporate actions and announcements.
    These are official filings, highly credible for citations.
    """
    
    BASE_URL = "https://www.nseindia.com"
    ANNOUNCEMENTS_API = f"{BASE_URL}/api/corporate-announcements"
    TIMEOUT_SECONDS = 10
    
    async def fetch_announcements(
        self,
        symbol: str,
        days_back: int = 7,
        max_results: int = 5
    ) -> List[Dict[str, Any]]:
        """
        Fetch corporate announcements from NSE
        
        Note: This is a placeholder implementation. In production:
        - Use NSE's official API with proper authentication
        - Handle rate limiting and API keys
        - Implement retry logic
        
        Args:
            symbol: NSE symbol (e.g., "RELIANCE", "TCS")
            days_back: How many days back to search
            max_results: Maximum announcements to return
        
        Returns:
            List of announcements with title, url, published_at
        """
        announcements = []
        
        try:
            # Clean symbol (remove .NS suffix)
            clean_symbol = symbol.replace('.NS', '').upper()
            
            # For now, return empty list with logging
            # TODO: Implement actual NSE API integration with proper auth
            logger.info(f"NSE announcements fetch not yet implemented for {clean_symbol}")
            
            # Placeholder: Would fetch from NSE API here
            # Note: NSE API often requires cookies, headers, and may have CAPTCHA
            # Consider using NSE's official data feed or third-party aggregators
        
        except Exception as e:
            logger.debug(f"NSE fetch failed: {e}")
        
        return announcements


class BSEAnnouncementsFetcher:
    """
    Fetch corporate announcements from BSE India
    
    Scrapes BSE's public announcements page for corporate filings.
    """
    
    BASE_URL = "https://www.bseindia.com"
    ANNOUNCEMENTS_URL = f"{BASE_URL}/corporates/ann.aspx"
    TIMEOUT_SECONDS = 10
    
    async def fetch_announcements(
        self,
        scrip_code: Optional[str] = None,
        company_name: Optional[str] = None,
        max_results: int = 5
    ) -> List[Dict[str, Any]]:
        """
        Fetch corporate announcements from BSE
        
        Args:
            scrip_code: BSE scrip code (optional)
            company_name: Company name for search (optional)
            max_results: Maximum announcements to return
        
        Returns:
            List of announcements with title, url, published_at
        """
        announcements = []
        
        try:
            # If we have .BO ticker, extract company name for search
            search_term = company_name if company_name else scrip_code
            
            async with httpx.AsyncClient(timeout=self.TIMEOUT_SECONDS) as client:
                logger.debug(f"Fetching BSE announcements for {search_term}")
                
                response = await client.get(
                    self.ANNOUNCEMENTS_URL,
                    headers={
                        'User-Agent': 'Mozilla/5.0 (Stock Intelligence Copilot)'
                    },
                    follow_redirects=True
                )
                
                if response.status_code != 200:
                    logger.warning(f"BSE returned status {response.status_code}")
                    return []
                
                soup = BeautifulSoup(response.text, 'html.parser')
                
                # Find announcement rows
                rows = soup.find_all('tr', class_='TTRow')
                
                for row in rows[:max_results]:
                    try:
                        cells = row.find_all('td')
                        if len(cells) < 4:
                            continue
                        
                        # Extract details (BSE table structure)
                        date_cell = cells[0].get_text(strip=True)
                        company_cell = cells[1].get_text(strip=True)
                        subject_cell = cells[2].get_text(strip=True)
                        
                        # Filter by company if specified
                        if search_term and search_term.lower() not in company_cell.lower():
                            continue
                        
                        # Check for PDF link
                        link_elem = cells[3].find('a')
                        pdf_url = ""
                        if link_elem and link_elem.get('href'):
                            pdf_url = f"{self.BASE_URL}{link_elem['href']}"
                        
                        # Parse date
                        published_at = None
                        if date_cell:
                            try:
                                published_at = datetime.strptime(date_cell, '%d %b %y')
                            except:
                                pass
                        
                        # Build announcement title
                        title = f"BSE: {subject_cell}"
                        if len(title) >= 10:
                            announcements.append({
                                'title': title,
                                'url': pdf_url if pdf_url else self.ANNOUNCEMENTS_URL,
                                'published_at': published_at,
                                'source': 'BSE India'
                            })
                    
                    except Exception as e:
                        logger.debug(f"Skipping BSE announcement: {e}")
                        continue
                
                logger.info(f"Fetched {len(announcements)} announcements from BSE")
        
        except httpx.TimeoutException:
            logger.warning("Timeout fetching from BSE")
        except Exception as e:
            logger.error(f"Error fetching BSE announcements: {e}")
        
        return announcements


# Helper function to aggregate news from all sources
async def fetch_all_indian_market_sources(
    ticker: str,
    company_name: Optional[str] = None,
    max_per_source: int = 3
) -> List[Dict[str, Any]]:
    """
    Fetch news/announcements from all Indian market sources
    
    Args:
        ticker: Stock ticker (e.g., "RELIANCE.NS", "TCS.BO")
        company_name: Company name for better search results
        max_per_source: Maximum items per source
    
    Returns:
        Combined list of news items from ET, NSE, BSE
    """
    all_news = []
    
    # Initialize fetchers
    et_fetcher = EconomicTimesMarketsFetcher()
    nse_fetcher = NSEAnnouncementsFetcher()
    bse_fetcher = BSEAnnouncementsFetcher()
    
    # Fetch from all sources concurrently
    try:
        results = await asyncio.gather(
            et_fetcher.fetch_stock_news(ticker, company_name, max_per_source),
            nse_fetcher.fetch_announcements(ticker, days_back=7, max_results=max_per_source),
            bse_fetcher.fetch_announcements(company_name=company_name, max_results=max_per_source),
            return_exceptions=True
        )
        
        # Combine results
        for result in results:
            if isinstance(result, list):
                all_news.extend(result)
            elif isinstance(result, Exception):
                logger.warning(f"Source fetch failed: {result}")
        
        logger.info(f"Total items fetched from all Indian sources: {len(all_news)}")
    
    except Exception as e:
        logger.error(f"Error aggregating Indian market sources: {e}")
    
    return all_news
